docs: fix filesystem boundary examples
